fixed send rlp tx

+++ lib/src/rust/api/transaction.dart
@@ -18,8 +18,8 @@ import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated.dart';
 // These function are ignored because they are on traits that is not defined in current crate (put an empty `#[frb]` on it to unignore): `fmt`
 
 Future<HistoricalTransactionInfo> sendSignedTransactions(
-        {required BigInt walletIndex,
-        required BigInt accountIndex,
+        {required int walletIndex,
+        required int accountIndex,
         required TransactionRequestInfo tx,
         required List<int> sig}) =>
     RustLib.instance.api.crateApiTransactionSendSignedTransactions(
